templates/python/eel/requirements.txt
eel==0.18.1,
pyinstaller==6.11.1,
python-dotenv==1.0.1,
